Добавил defer в /usercard, удалил класс Picture вместо него оставил просто списки. В методе on_member_remove добавил проверку на бота. Проработал коианду /gift. Теперь команда /balik называется /money и она успешно проработана как и команда /pay, добавил новый метод payment в класс DataBaseFuncs.

+++ bot/Cogs/members.py
@@ -1,10 +1,11 @@
+import random
 import disnake
 
 from datetime import datetime as dt
 from disnake.ext import commands
 
 from database.src.base import data_funcs
-from assets.pictures import Picture
+from assets.pictures import bot_images
 
 
 
@@ -12,7 +13,6 @@ from assets.pictures import Picture
 class MembersAction(commands.Cog):
      def __init__(self, bot: commands.Bot) -> None:
           self.bot = bot
-          self.picrure = Picture()
      
      
      @commands.Cog.listener()
@@ -24,7 +24,7 @@ class MembersAction(commands.Cog):
                colour=disnake.Colour.blue(),
                timestamp=dt.now()
           )
-          embed.set_thumbnail(file=disnake.File(fp=self.picrure.get_random_bot_images))
+          embed.set_thumbnail(file=disnake.File(fp=random.choice(bot_images)))
                     
           await member.send(embed=embed)
           
@@ -32,10 +32,11 @@ class MembersAction(commands.Cog):
           
      @commands.Cog.listener()
      async def on_member_remove(self, member: disnake.Member) -> None:
-          await data_funcs.delete_user(
-               id_guild=member.guild.id,
-               id_member=member.id
-          )
+          if not member.bot:
+               await data_funcs.delete_user(
+                    id_guild=member.guild.id,
+                    id_member=member.id
+               )
